Crate dbox

Dropbox SDK for Rust

This crate implements a simple client for the Dropbox API. It uses hyper for HTTP, though that part is swappable if you would like to use a different HTTP implementation.

If you want to use hyper, just use dbox::client::Client, as in the example below. If, however, you want to use something else, you will need to implement the DropboxClient trait for your data structure, build this crate with --no-default-features, and pass a reference to your client as the first parameter to the API functions.

Example

extern crate dbox;

use std::env;
use dbox::client::Client;
use dbox::files;

let access_token = env::var("DROPBOX_TOKEN");
let client = Client::new(access_token);
let folder_list = files::copy_(&client, "/path/to/existing/file", "/path/to/new/file");

Modules

client
Default implementation of the Dropbox Client

files
Module for performing operations on files in a dropbox account

sharing
TODO

structs
Module that holds definitions for dropbox data structures

users
TODO

Structs

Response
Simple abstraction of a HTTP response, to allow the HTTP client to be pluggable

Enums

ApiError
Collection of possible errors

Traits

DropboxClient

Type Definitions

Result