Module dbox::files

Module for performing operations on files in a dropbox account

Structs

CommitInfo

FolderListLongpoll

GetCursorOptions

ListRevisions

Search

SearchMatch

SearchOptions

ThumbnailOptions

UploadOptions
Optional arguments to the upload API call

UploadSessionCursor

Enums

SearchMatchType

SearchMode

ThumbnailFormat

ThumbnailSize

WriteMode
Instructs dropbox what to do when a conflict happens during upload

Functions

copy_
Copy a file

create_folder
Create a folder

delete
Delete a file or folder from the user's dropbox acconut

download
Download a file

download_to

download_to_file
TODO implement

get_metadata
Get metadata for a file

get_preview
TODO implement

get_preview_to_file
TODO implement

get_thumbnail

get_thumbnail_to_file

get_thumbnail_to_file_with_options
TODO implement

get_thumbnail_with_options
TODO implement

list_folder
List the entries in a user's dropbox folder

list_folder_continue
TODO implement

list_folder_get_latest_cursor

list_folder_get_latest_cursor_with_options
TODO implement

list_folder_longpoll
TODO implement

list_revisions
TODO implement

move_
Move a file in a user's dropbox

permanently_delete
TODO implement

restore
TODO implement

search

search_with_options
TODO implement

upload
Upload a file to the user's dropbox acconut

upload_session_append
TODO implement

upload_session_finish
TODO implement

upload_session_start
TODO implement

upload_with_options
Upload a file to the user's dropbox acconut